Implement some decent scaling
Replace the naive nearest-neighbor scaling with a much better method.
This is based on the "high quality" scaling filter I implemented in
OpenBOR a couple of years ago. It's equivalent to scaling with nearest
neighbor to above the desired size, then downsampling with bilinear
filtering to the actual desired size. And at integer multiples, it's
completely identical to nearest neighbor.