Add support for closing games and opening new ones afterwards.
In theory, calling retro_unload_game() and retro_deinit() should be
enough to clear everything for a new game to be loaded. In practice,
the core doesn't play any audio in games loaded after the first. This is
solved by completely unloading the core's .so from memory and reloading
it when opening another game.