Reduce audio lag and fix A/V sync after audio interruption
Reducing the audio lag a bit is as easy as reducing the buffer size from
4096 samples (~5.6 frames) to 1596 (~2.2 frames).

A/V sync is a thornier issue, but for now, it can correct for major
interruptions of the emulator thread. One such interruption can easily
be created by displaying the "open" dialog for the first time while the
emulation thread is running.