Make texture size equal to current PC Engine output resolution.
Instead of having a texture of max_width*max_height and updating its top
left corner. This allows the texture wrap setting of GL_CLAMP_TO_EDGE to
apply to the bottom and right of the texture, not just the top and left.
Fixes white lines along the right and bottom edges of the display area.